Buffer refactor. Probably new errors, write tests soon.
